ULTI-419032 - Alex - refactored WithTimeout
